+++ src/ba_timeseries_gradients/utils.py
@@ -46,9 +46,6 @@ def save_hdf5(
     with h5py.File(filename, "w") as h5_file:
         h5_file.create_dataset("gradients", data=output_gradients)
         h5_file.create_dataset("lambdas", data=lambdas)
-    with h5py.File(filename, "w") as h5_file:
-        h5_file.create_dataset("gradients", data=output_gradients)
-        h5_file.create_dataset("lambdas", data=lambdas)
 
 
 def save_json(
